Use different event and move to the appropriate class
